🩹 change interface CLPool

+++ src/asset-modules/Slipstream/interfaces/ICLPool.sol
@@ -23,7 +23,7 @@ interface ICLPool {
             int128 stakedLiquidityNet,
             uint256 feeGrowthOutside0X128,
             uint256 feeGrowthOutside1X128,
-            uint128 rewardGrowthOutsideX128,
+            uint256 rewardGrowthOutsideX128,
             int56 tickCumulativeOutside,
             uint160 secondsPerLiquidityOutsideX128,
             uint32 secondsOutside,
